+++ ninja-core/src/main/java/ninja/bodyparser/BodyParserEngineMultipartPost.java
@@ -13,6 +13,7 @@ public class BodyParserEngineMultipartPost extends BodyParserEnginePost {
         super(paramParsers);
     }
 
+    @Override
     public String getContentType() {
         return "multipart/form-data";
     }