+++ ninja-servlet-integration-test/src/main/java/models/PersonWithParameters.java
@@ -1,0 +1,44 @@
+package models;
+
+import java.util.Map;
+
+public class PersonWithParameters {
+ 
+    private String name;
+    private Map<String, String> parameters;
+    private Long a;
+    private String b;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public Map<String, String> getParameters() {
+        return parameters;
+    }
+
+    public void setParameters(Map<String, String> parameters) {
+        this.parameters = parameters;
+    }
+
+    public Long getA() {
+        return a;
+    }
+
+    public void setA(Long a) {
+        this.a = a;
+    }
+
+    public String getB() {
+        return b;
+    }
+
+    public void setB(String b) {
+        this.b = b;
+    }
+    
+}